chore: moving and renaming utils for extract_metrics
