+++ src/textdescriptives/extractors.py
@@ -9,6 +9,8 @@ from spacy.tokens import Doc, Span, Token
 from wasabi import msg
 
 from textdescriptives.utils import (
+    _create_spacy_pipeline,
+    _remove_textdescriptives_extensions,
     get_doc_assigns,
     get_span_assigns,
     get_token_assigns,
@@ -109,108 +111,11 @@ def extract_df(
     return pd.DataFrame(extract_dict(docs, metrics, include_text))
 
 
-def download_spacy_model(lang: str, size: str) -> str:
-    """Download a large spacy model for a given language.
-
-    Args:
-        lang (str): Language to download a model for.
-        size (str): Size of the model to download. One of "sm", "md", "lg", "trf".
-
-    Returns:
-        str: Name of the downloaded model.
-    """
-    data_source = "news" if lang != "en" else "web"
-    spacy_model = f"{lang}_core_{data_source}_{size}"
-    # don't download if model already exists
-    if spacy_model in spacy.cli.info()["pipelines"]:  # type: ignore
-        return spacy_model
-    spacy.cli.download(spacy_model)
-    return spacy_model
-
-
-def load_spacy_model(
-    spacy_model: Optional[str],
-    lang: Optional[str],
-    metrics: Iterable[str],
-    spacy_model_size: str,
-) -> Language:
-    """Load a spacy model suitable for the metrics to be extracted.
-
-    Args:
-        spacy_model (Optional[str]): The spacy model to use. If not set,
-            will download one based on lang.
-        lang (Optional[str]): Language of the text. If lang is set and no spacy
-            model is provided, will automatically download and use a spacy
-            model for the language.
-        metrics (Iterable[str]): Which metrics to extract.
-        spacy_model_size (str): Size of the spacy model to download.
-
-    Returns:
-        Language: a spacy pipeline
-    """
-
-    metrics_requiring_spacy_model = {
-        "dependency_distance",
-        "pos_stats",
-        "coherence",
-        "pos_proportions",
-    }
-    # if no spacy model is necessary for the metrics, return a blank model
-    # for the language
-    if not bool(metrics_requiring_spacy_model.intersection(metrics)):
-        if lang is not None:
-            return spacy.blank(lang)
-        if spacy_model:
-            lang = spacy_model.split("_")[0]
-            return spacy.blank(lang)
-        raise ValueError(
-            "No spacy model needed for the metrics and no language provided. "
-            + "Please provide a language to create a blank spacy model.",
-        )
-    # if no spacy model is provided, download one
-    if spacy_model is None:
-        if lang is None:
-            raise ValueError(
-                "No spacy model provided and no language provided. "
-                + "Please provide a language to download a spacy model.",
-            )
-        msg.info(f"No spacy model provided. Inferring spacy model for {lang}.")
-        spacy_model = download_spacy_model(lang=lang, size=spacy_model_size)
-    return spacy.load(spacy_model)
-
-
-def _remove_spacy_extension(
-    language: Union[Type[Doc], Type[Span], Type[Token]],
-    extension: str,
-) -> None:
-    """Remove spacy extension from a Language object if it exists."""
-    if language.has_extension(extension):
-        language.remove_extension(extension)
-
-
-def _remove_textdescriptives_extensions() -> None:
-    """Remove spacy extensions added by textdescriptives.
-
-    This is necessary to avoid errors if running `extract_metrics`
-    multiple times with different metrics
-    """
-    for metric in get_valid_metrics():
-        doc_assigns = get_doc_assigns(metric)
-        for assigned in doc_assigns:
-            _remove_spacy_extension(language=Doc, extension=assigned)
-        span_assigns = get_span_assigns(metric)
-        for assigned in span_assigns:
-            _remove_spacy_extension(language=Span, extension=assigned)
-        token_assings = get_token_assigns(metric)
-        for assigned in token_assings:
-            _remove_spacy_extension(language=Token, extension=assigned)
-
-
 def extract_metrics(
     text: Union[str, List[str]],
-    spacy_model: Optional[str] = None,
     lang: Optional[str] = None,
     metrics: Optional[Iterable[str]] = None,
+    spacy_model: Optional[str] = None,
     spacy_model_size: str = "lg",
 ) -> pd.DataFrame:
     """Extract metrics from a text or a list of texts to a Pandas dataframe.
@@ -218,9 +123,7 @@ def extract_metrics(
     Args:
         text (Union[str, List[str]]): A text or a list of texts.
         spacy_model (str, optional): The spacy model to use. If not set,
-            will download one based on lang. We recommend always using `spacy_model`
-            and downloading the model beforehand to have more control over the
-            model as well as to avoid versioning issues. Defaults to None.
+            will download one based on lang. Defaults to None.
         lang (str, optional): Language of the text. If lang is set and no spacy
             model is provided, will automatically download and use a spacy
             model for the language. Defaults to None.
@@ -239,6 +142,12 @@ def extract_metrics(
     if spacy_model is None and lang is None:
         raise ValueError("Either a spacy model or a language must be provided.")
 
+    if spacy_model is not None and lang is not None:
+        msg.info(
+            "Both a spacy model and a language were provided. "
+            + "Will use the spacy model and ignore language."
+        )
+
     if metrics is None:
         metrics = get_valid_metrics()
 
@@ -246,7 +155,7 @@ def extract_metrics(
     _remove_textdescriptives_extensions()
 
     # load spacy model if any component requires it
-    nlp = load_spacy_model(
+    nlp = _create_spacy_pipeline(
         spacy_model=spacy_model,
         lang=lang,
         metrics=metrics,
